feat: add additional params to pkp-ether
